# Always print explanation for univ inconsistency, rm Flags.univ_print