--- a/lib/flags.ml
+++ b/lib/flags.ml
@@ -60,7 +60,6 @@ let profile = false
 let ide_slave = ref false
 
 let raw_print = ref false
-let univ_print = ref false
 
 let we_are_parsing = ref false