@ejgallego 2018-04-25
> The lazyness is because getting an explanation may be costly and we
> may discard it without printing.

who's getting an explanation and not printing it?