@ejgallego 2018-04-25
> who's getting an explanation and not printing it?

It is the tactic layer? If so, we should absolutely fix that.