@SkySkimmer 2018-04-25
>who's getting an explanation and not printing it?

Any tactic like `foo || bar` where foo fails because of a universe inconsistency.
`Reductionops.infer_conv` when syntactic comparison fails because of an inconsistency.
etc