@ejgallego 2018-04-25
Ah, somehow I would have expected that `univ_inconsistency` would be:
```ocaml
type univ_inconsistency = constraint_type * Universe.t * Universe.t * UGraph.t
```
and that `explain_universe_inconsistency` would actually call the "explainer"