@ejgallego 2018-04-25
`UGraph.t` is a functional structure, right? If so, then I'd prefer to embed pointers in exceptions instead of elaborated terms [either under a closure or not]