@SkySkimmer 2018-04-25
`type univ_inconsistency` is defined in the wrong module for that.