@SkySkimmer 2018-04-26
>Indeed the errors look conceptually different to me

What do you mean?

>Also, what do you think about fixing the weird call in univdecls?

Some other PR can do that.