@ejgallego 2018-04-26
I mean that the errors happen in different contexts/algorithms, so it could make sense to separate them.

So do you want this to go as in?