@SkySkimmer 2018-04-26
There's a short path for an error on `Prop = Set` and an error on `u+k+1 <= u` (k>=1).
The both seem like they could result in uncaught exceptions since the user may not expect `Constraint.t -> Constraint.t` functions to do constraint checking.
I don't know how the second can happen. The coverage is interesting btw https://x80.org/coq2017/report/file0247.html#L658 the anomaly happens a few times (!!) but not the special case.