@ejgallego 2018-04-26
Sounds good, given that this seems to be impacting performance it is IMHO worth to try to allocate only the exception instead of the whole lazy closure.